requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
asgiref==3.5.0
autopep8==2.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
idna==3.3
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
